# docs/openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/tokens/{chainId}": {
      "get": {
        "operationId": "TokensController_getTokens",
        "parameters": [
          {
            "name": "chainId",
            "required": true,
            "in": "path",
            "description": "Chain ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns an array of tokens",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TokenDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all tokens for a specific chain",
        "tags": [
          "tokens"
        ]
      }
    },
    "/tokens/{chainId}/{address}": {
      "get": {
        "operationId": "TokensController_getToken",
        "parameters": [
          {
            "name": "chainId",
            "required": true,
            "in": "path",
            "description": "Chain ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "address",
            "required": true,
            "in": "path",
            "description": "Token address",
            "schema": {
              "example": "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns token information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenDto"
                }
              }
            }
          },
          "404": {
            "description": "Token not found"
          }
        },
        "summary": "Get a specific token by address",
        "tags": [
          "tokens"
        ]
      }
    },
    "/prices/{chainId}": {
      "get": {
        "operationId": "PricesController_getPrices",
        "parameters": [
          {
            "name": "chainId",
            "required": true,
            "in": "path",
            "description": "Chain ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns all token prices",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get all token prices for a specific chain",
        "tags": [
          "prices"
        ]
      }
    },
    "/prices/{chainId}/{address}": {
      "get": {
        "operationId": "PricesController_getPrice",
        "parameters": [
          {
            "name": "chainId",
            "required": true,
            "in": "path",
            "description": "Chain ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "address",
            "required": true,
            "in": "path",
            "description": "Token address",
            "schema": {
              "example": "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns token price information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPriceDto"
                }
              }
            }
          },
          "404": {
            "description": "Token price not found"
          }
        },
        "summary": "Get price for a specific token",
        "tags": [
          "prices"
        ]
      }
    },
    "/swaps/quote/{chainId}": {
      "get": {
        "operationId": "SwapsController_getQuote",
        "parameters": [
          {
            "name": "chainId",
            "required": true,
            "in": "path",
            "description": "Chain ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "fromToken",
            "required": true,
            "in": "query",
            "description": "Source token address",
            "schema": {
              "example": "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48",
              "type": "string"
            }
          },
          {
            "name": "toToken",
            "required": true,
            "in": "query",
            "description": "Destination token address",
            "schema": {
              "example": "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2",
              "type": "string"
            }
          },
          {
            "name": "amount",
            "required": true,
            "in": "query",
            "description": "Amount of source token to swap (in wei/lamports)",
            "schema": {
              "example": "1000000000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns swap quote information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwapQuoteDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters"
          },
          "404": {
            "description": "No routes found"
          }
        },
        "summary": "Get swap quote",
        "tags": [
          "swaps"
        ]
      }
    }
  },
  "info": {
    "title": "Dextract-fi API",
    "description": "The Dextract-fi API documentation",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "TokenDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "description": "Token address"
          },
          "symbol": {
            "type": "string",
            "description": "Token symbol"
          },
          "name": {
            "type": "string",
            "description": "Token name"
          },
          "decimals": {
            "type": "number",
            "description": "Token decimals"
          },
          "logoURI": {
            "type": "string",
            "description": "Token logo URI"
          },
          "tags": {
            "description": "Token tags",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "chainId": {
            "type": "number",
            "description": "Chain ID of the token"
          }
        },
        "required": [
          "address",
          "symbol",
          "name",
          "decimals",
          "chainId"
        ]
      },
      "PriceResponseDto": {
        "type": "object",
        "properties": {
          "prices": {
            "type": "object",
            "description": "Map of token addresses to their price information",
            "additionalProperties": {
              "type": "object",
              "$ref": "#/components/schemas/TokenPriceDto"
            }
          },
          "updatedAt": {
            "type": "number",
            "description": "Timestamp when prices were last updated"
          }
        },
        "required": [
          "prices",
          "updatedAt"
        ]
      },
      "TokenPriceDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "description": "Token address"
          },
          "priceUsd": {
            "type": "number",
            "description": "Token price in USD"
          },
          "timestamp": {
            "type": "number",
            "description": "Price timestamp"
          },
          "change24h": {
            "type": "number",
            "description": "24-hour price change percentage"
          },
          "change7d": {
            "type": "number",
            "description": "7-day price change percentage"
          },
          "volume24h": {
            "type": "number",
            "description": "24-hour trading volume in USD"
          },
          "marketCap": {
            "type": "number",
            "description": "Token market capitalization in USD"
          }
        },
        "required": [
          "address",
          "priceUsd",
          "timestamp"
        ]
      },
      "SwapRouteDto": {
        "type": "object",
        "properties": {
          "fromToken": {
            "type": "string",
            "description": "Source token address"
          },
          "toToken": {
            "type": "string",
            "description": "Destination token address"
          },
          "fromAmount": {
            "type": "string",
            "description": "Amount of source token to swap (in wei/lamports)"
          },
          "toAmount": {
            "type": "string",
            "description": "Expected amount of destination token to receive (in wei/lamports)"
          },
          "priceImpact": {
            "type": "number",
            "description": "Price impact of the swap as a percentage"
          },
          "path": {
            "description": "Token addresses in the swap path",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "providers": {
            "description": "DEX providers used in the route",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "estimatedGas": {
            "type": "string",
            "description": "Estimated gas cost (in wei/lamports)"
          }
        },
        "required": [
          "fromToken",
          "toToken",
          "fromAmount",
          "toAmount",
          "priceImpact",
          "path",
          "providers"
        ]
      },
      "SwapQuoteDto": {
        "type": "object",
        "properties": {
          "routes": {
            "description": "All available swap routes",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SwapRouteDto"
            }
          },
          "bestRoute": {
            "description": "Best swap route based on output amount",
            "allOf": [
              {
                "$ref": "#/components/schemas/SwapRouteDto"
              }
            ]
          },
          "fromToken": {
            "type": "string",
            "description": "Source token address"
          },
          "toToken": {
            "type": "string",
            "description": "Destination token address"
          },
          "fromAmount": {
            "type": "string",
            "description": "Amount of source token to swap (in wei/lamports)"
          },
          "updatedAt": {
            "type": "number",
            "description": "Timestamp when the quote was generated"
          }
        },
        "required": [
          "routes",
          "bestRoute",
          "fromToken",
          "toToken",
          "fromAmount",
          "updatedAt"
        ]
      }
    }
  }
}